[backend] refactor: change dir name
